backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
motor
slowapi
passlib[bcrypt]
python-multipart
python-jose[cryptography]
authlib
itsdangerous
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-pinecone
pinecone-client
tiktoken
pymupdf
presidio-analyzer
presidio-anonymizer
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
pybreaker
langfuse
upstash-redis
